Feature/18 vertexedge graph io managment (#50)
* vertex-edge graph reader V1 - before human check

* Fix the logger  use of logger handler

* Changed graph vertex colors to be uint32 instead of int32

* Changed graph vertex colors to be uint32 instead of int32

* formated graph utils

* Moved the io files into a graph folder inside io

* Fixed vertex edge reader to read uint colors

* Updated cmakes

* Formatted code

* Updated vertex edge reader - no line indices

* Written tests

* Reformatted vertex edge reader

* moved exception handeling to the most specific place

* Cleaned up vertex-edge graph reader

* remove sorting of vertices before remapping

* Updated tests

* Fixed gettng src and dst vertrex beore checking for color

* fix test cases with no definitive rema[pig order

* Added base laggable class file

* Inheritance logging solution changed to wrapper class

* Updated logger handler to have is null and right constructor name

* Updated LoggerHandler to have is null function

* Using the logger in readers

* Using the logger in  readers implemintation

* Updated tests to match

* Fix the logger  use of logger handler

* Moved the io files into a graph folder inside io

* Changed graph vertex colors to be uint32 instead of int32

* Fixed vertex edge reader to read uint colors

* Updated cmakes

* Formatted code

* Updated vertex edge reader - no line indices

* Written tests

* Reformatted vertex edge reader

* moved exception handeling to the most specific place

* Cleaned up vertex-edge graph reader

* remove sorting of vertices before remapping

* Updated tests

* Fixed gettng src and dst vertrex beore checking for color

* fix test cases with no definitive rema[pig order

* updated clang tidy

* done clang-format

* Fixed linting errors

* Updated clang tools version to specific one in enviorment yml

* clang-formatted code

* Fixed clang tidy erros

* Updated ci to use clang-tidy 17

* Returned correct initializer to coloredgraphtests and quieted misslining linter problem

* Fixed formatting

* Reduce scope of clang-tidy in ci

* Updated the clang-tidy command in claude.md

* Updated consts and basic class functions definitions in colored graph and readers

* Updated naming conventions in logger handler

* Added nodiscard to exceptions

* Updated uint convention in constants

* Updated clang-tidy rules

+++ include/utils/GraphUtils.h
@@ -124,8 +124,8 @@ private:
      * @throws GraphConstructionException if any color value exceeds MAX_VERTEX_COLOR.
      */
     template <typename GraphType, typename GetColor>
-    static std::vector<int32_t> extract_vertex_colors(const GraphType& boost_graph,
-                                                      const GetColor& get_color);
+    static std::vector<uint32_t> extract_vertex_colors(const GraphType& boost_graph,
+                                                       const GetColor& get_color);
 
     /**
      * @brief Extracts colored edges using a caller-supplied color accessor.
@@ -158,7 +158,7 @@ GraphUtils::convert_boost_graph_to_colored_graph(const GraphType& boost_graph,
                                                  std::map<std::string, uint32_t>& color_map)
 {
     const uint32_t num_vertices = static_cast<uint32_t>(boost::num_vertices(boost_graph));
-    const std::vector<int32_t> vertex_colors = extract_vertex_colors(
+    const std::vector<uint32_t> vertex_colors = extract_vertex_colors(
         boost_graph,
         [&](const typename boost::graph_traits<GraphType>::vertex_descriptor& vertex_desc)
         {
@@ -206,10 +206,10 @@ GraphUtils::extract_uncolored_edges(const GraphType& boost_graph)
 }
 
 template <typename GraphType, typename GetColor>
-std::vector<int32_t> GraphUtils::extract_vertex_colors(const GraphType& boost_graph,
-                                                       const GetColor& get_color)
+std::vector<uint32_t> GraphUtils::extract_vertex_colors(const GraphType& boost_graph,
+                                                        const GetColor& get_color)
 {
-    std::vector<int32_t> vertex_colors;
+    std::vector<uint32_t> vertex_colors;
     vertex_colors.reserve(boost::num_vertices(boost_graph));
     for (const auto& vertex_desc : boost::make_iterator_range(boost::vertices(boost_graph)))
     {
@@ -219,7 +219,7 @@ std::vector<int32_t> GraphUtils::extract_vertex_colors(const GraphType& boost_gr
             throw GraphConstructionException("vertex color " + std::to_string(color) +
                                              " exceeds maximum allowed value");
         }
-        vertex_colors.push_back(static_cast<int32_t>(color));
+        vertex_colors.push_back(color);
     }
     return vertex_colors;
 }
